+++ app/Repository/CampaignRepository.php
@@ -106,6 +106,9 @@ class CampaignRepository
      */
     public function save(Campaign $campaign, array $banners = [], array $conversions = []): Campaign
     {
+        if (Campaign::STATUS_ACTIVE === $campaign->status && empty($banners)) {
+            throw new InvalidArgumentException('Cannot save active campaign without creatives');
+        }
         DB::beginTransaction();
         $status = $campaign->status;
         $campaign->status = Campaign::STATUS_DRAFT;
@@ -236,6 +239,13 @@ class CampaignRepository
 
             $this->bannerClassificationCreator->createForCampaign($campaign);
             $campaign->refresh();
+
+            if (
+                Campaign::STATUS_ACTIVE === $campaign->status
+                && !$campaign->banners()->where('status', Banner::STATUS_ACTIVE)->exists()
+            ) {
+                throw new InvalidArgumentException('Cannot update active campaign without creatives');
+            }
             DB::commit();
         } catch (InvalidArgumentException $exception) {
             DB::rollBack();