+++ CHANGELOG.md
@@ -7,8 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Do not allow to update active campaign without creatives
 ### Removed
 - Redundant pagination data
+### Fixed
+- Allow creating campaign draft without creatives
 
 ## [2.2.3] - 2022-12-19
 ### Added